Added in voting rewards
Added in a method to define reward packages that cost voting points.
Any function starting with VOTEREWARD_ are classed as packages, and can
be called whenever a user requests a package. e.g. VOTEREWARD_CREDITS
will be called whenever someone enters CREDITS as the package.
votes_deduct is used to check if the user has enough voting points for
the package, and deducts the points from the users "account"
chatcommand_VOTEPOINTS just tells the user how many voting points they
have to spend