Fixes and rewrites
Sorted and fixed the lib_check_server_running so it works properly with
all the commands now.  Soft-restart was added and chatcommand restart
now uses it.  I re-wrote resultloop to be even faster and very simple.
There is still a bug that will cause some chatcommands to not register.
This will likely require a re-write of the search functions to fix .