Ram drive and sleep to logloop
Spent some time watching the script run, it was using far to much CPU so
I switched all file writes to the ram drive and also added a sleep 1
into the log loop file.  This does change the responsiveness slightly,
and can be adjusted if needed.  This took CPU usage from a stead 7% to
1% on my test rig