Added credit limit and use of !deposit/withdraw all
Added a credit limit that is specified in the config file with a numerical value. Also added the ability to use !deposit/withdraw all to deposit all of the players current credits if the limit is off or as much credits as they can deposit until they hit the credit limit if it's turned on.

Need to add $CONFIGDTSD_CREDITLIMITON and $CONFIGDTSD_CREDITLIMIT to the config file and add them to the setup file.