Testing new functions for core loops
First successful test to replace existing log loops.  The new system
tails the log into mysql, then searches log as lines are fed, and
finally executes these commands from the database.  This new function
uses less CPU and is quicker and cleaner plus it will allow commands to
be inserted into the database to be ran from the script which has also
been tested and works.