Testing Server Check
Server check is ran at 5 minute intervals for default setting and looks
to see if the server is running, if not it starts it.  If the server is
running it send a message and then looks for it in the log.  If it finds
it then it stands to reason the server is responding.  Right now it just
waits 10 seconds and checks to see, if the server is unresponsive it
gets a restart.  You will want to disable this for server tasks if you
don't want the server to keep launching with shadow.dtsd clearschedule
which will stop all jobs until shadow is ran again.