Bug fixes, timed sector updates
Fix for name formatting on boarding commands related to different string
for authenticating servers.  Sector update only triggered every 10
minutes on sector change and login now due to spammy action of a lot of
people on large server with a lot of entities.  Fix for voting bug
related to no string expansion for variable.