Testing Duel, bugs, troubleshooting
Changed logloop to sleep only if no new lines are detected.  This will
speed it up when needed without taxing the CPU.  Added new lib for mysql
that will be used to replace simplify most commands during the next
rewrite.  Moved the java execution lines to defaults.dtsd so it can be
edited easily.  Added check to make sure player is online in the
database for dueling.   Added in functions for player upon sector change
without ship.  Further testing of duel with first successful test of
most basic functions.  Added in trigger commands to return player to
sector upon dying or return or give a ship to a player.  Added more
verbose message and redundancy check into kill.  Fixed bug with dump
command related to paths for mysql tables.  Set the logs to be dumped
upon every restart into the dump directory with a timestamp of the dump
time.